// dist/es2015/renderer/shader/sourceBuilder/ShaderSourceBuilder.d.ts
/// <reference types="wonder-commonlib" />
import { Hash } from "wonder-commonlib/dist/es2015/Hash";
import { ShaderData } from "../shader/Shader";
export declare abstract class ShaderSourceBuilder {
    attributes: Hash<ShaderData>;
    uniforms: Hash<ShaderData>;
    vsSource: string;
    fsSource: string;
    abstract build(...args: any[]): void;
    abstract clearShaderDefinition(): void;
    dispose(): void;
    protected convertAttributesData(): void;
}
export declare type SourceDefine = {
    name: string;
    value: any;
};